project/requirements.txt
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
django-cors-headers==4.7.0
newsapi-python==0.2.7
numpy==1.24.1
fasttext==0.9.3
psycopg2-binary>=2.9.10
uvicorn==0.40.0
yake==0.6.0